LayoutTests/fast/css-intrinsic-dimensions/resize-inside-percent-width-overflow-hidden.html

<!DOCTYPE html>
<p>There should be a blue square below.</p>
<div id="root" style="width:fit-content;" data-expected-width="100">
    <div style="width:100%; height:10em; overflow:hidden;" data-expected-width="100">
        <div id="child" style="width:20px; height:100px; background:blue;"></div>
    </div>
</div>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<script>
    document.body.offsetTop;
    document.getElementById("child").style.width = "100px";
    checkLayout("#root");
</script>
